Update
Finally finished the project. I worked on the third and fourth section,
plus the footer.

The only part with which I am not satisfied with is the header. A part there
doesn't look exactly like the original part.

So the next time I come back to this project, the first thing I should probably
work on should be the header.